requirements.txt
torch==2.0.1
torchvision==0.15.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.1
numpy==1.21.6
pandas==2.0.3
Pillow==9.4.0
imageio==2.26.0
scikit-image==0.19.3